sdk: split query_verbs into module directory (2/9: extract test module)
Move mod tests body and mod fingerprint_bench to query_verbs/tests.rs;
replace inline block with #[cfg(test)] mod tests; declaration.
